\name{brycesite}
\docType{data}
\alias{brycesite}
\title{Site Data for Bryce Canyon National Park}
\description{Environmental variables recorded at or calculated for each of 160
sample plots in Bryce Canyon National Park, Utah, U.S.}
\usage{data(brycesite)}
\format{a data.frame with plots as rows and site variables as columns.
Variables are:
\describe{
\item{labels}{original plot codes}
\item{annrad}{annual direct solar radiation in Langleys}
\item{asp}{slope aspect in degrees}
\item{av}{aspect value = (1+cosd(asp-30)/2)}
\item{depth}{soil depth = "deep" or "shallow"}
\item{east}{UTM easting}
\item{elev}{elevation in feet}
\item{grorad}{growing season radiation in Langleys}
\item{north}{UTM northing}
\item{pos}{topographic position}
\item{quad}{USGS 7.5 minute quad sheet}
\item{slope}{percent slope}
}
}

\keyword{datasets}
