\name{labdsv.internal}
\alias{plotid}
\alias{specid}
\alias{surf}
\alias{jsurf}
\alias{hilight}
\alias{chullord}
\alias{orddist}
\alias{shannon}
\alias{pip}
\title{LabDSV Internal Functions}
\description{These functions establish several generic functions,
and are not intended to be called directly}
\keyword{internal}
