\name{bryceveg}
\docType{data}
\alias{bryceveg}
\title{Bryce Canyon Vegetation Data}
\description{Estimates of cover class for all non-tree vascular plant 
species in 160 375\eqn{m^2}
circular sample plots. Species codes are first three letters of genus + 
first three letters of specific epithet.} 
\usage{data(bryceveg)}
\format{a data.frame of 160 sample units (rows) and 169 species (columns).
Cover is estimated in codes as follows:
\describe{
\item{0.2}{present in the stand but not the plot}
\item{0.5}{0-1\%}
\item{1.0}{1-5\%}
\item{2.0}{5-25\%}
\item{3.0}{25-50\%}
\item{4.0}{50-75\%}
\item{5.0}{75-95\%}
\item{6.0}{95-100\%}
}
}
\keyword{datasets}
